add changes that I could from PR comments
