+++ app/assets/stylesheets/application.scss
@@ -1,1 +1,5 @@
 @import "bootstrap";
+
+.navbar {
+    background-color: #e3f2fd;
+}
