+++ app/controllers/searches_controller.rb
@@ -11,56 +11,6 @@ class SearchesController < ApplicationController
       end
 
       render :index
-      # else
-      #   flash.now[:alert] = 'Please enter a city name and select a type to search'
-      #   @places = []
-      #   # render :index
     end
   end
-
-  def add_to_favorites
-    place_id = params[:place][:id].to_i
-    place = params[:place][:place_type].singularize
-
-    case place
-    when "Attraction"
-      if current_user.attraction_ids.include?(place_id)
-        flash[:alert] = failure_message(place)
-      else
-        current_user.attractions << Attraction.find(place_id)
-        flash[:notice] = success_message(place)
-      end
-    when "Hotel"
-      if current_user.hotel_ids.include?(place_id)
-        flash[:alert] = failure_message(place)
-      else
-        current_user.hotels << Hotel.find(place_id)
-        flash[:notice] = success_message(place_id)
-      end
-    when "Restaurant"
-      if current_user.restaurant_ids.include?(place_id)
-        flash[:alert] = failure_message(place)
-      else
-        current_user.hotels << Restaurant.find(place_id)
-        flash[:notice] = success_message(place)
-      end
-    else
-      raise PlaceTypeInvalid.new("#{place} is not supported at the moment")
-    end
-
-    redirect_to search_path
-  end
-
-  private
-    def place_params
-      params.require[:place].permit(:id, :place_type)
-    end
-
-    def failure_message(place_type)
-      "The #{place_type} has already been added please select another one"
-    end
-
-    def success_message(place_type)
-      "The #{place_type} was added successfully"
-    end
 end
