+++ app/helpers/places_helper.rb
@@ -1,0 +1,2 @@
+module PlacesHelper
+end
