+++ spec/helpers/places_helper_spec.rb
@@ -1,0 +1,14 @@
+require 'rails_helper'
+
+# Specs in this file have access to a helper object that includes
+# the PlacesHelper. For example:
+#
+# describe PlacesHelper do
+#   describe "string concat" do
+#     it "concats two strings with spaces" do
+#       expect(helper.concat_strings("this","that")).to eq("this that")
+#     end
+#   end
+# end
+RSpec.describe PlacesHelper, type: :helper do
+end
