+++ spec/requests/places_spec.rb
@@ -1,0 +1,4 @@
+require 'rails_helper'
+
+RSpec.describe "Places", type: :request do
+end
